Refactored Animation, Controller & added interface
Refactored Animation so it can be used for buildings as well as tiles.
It now uses the FrameRefresher.

Refactored Controller into a ControlMapper which can register a function
to be executed for a given key down/up/pressed.

Added canvases for interface and a basic class for it. Interface is in
very early stage at this point.

Engine now also update()s and draw()s the interface in its step.

Added a simple .css to help with canvases' styles.