Added BuildingPanel and PlayerBoardController
Added a BuildingPanel which can be used to select a building and place
it on the board. Building appears and follows the mouse inside the
gameWorld canvas until the player clicks. Still need to implement the
creation of the building on the click as well as locking the build
placement to regions.