Added Ant Trails & Instructions
Game is finalized. Player can now connect buildings using AntTrails to
move resources between them. Added an Instructions button which pops ups
a jquery-ui dialog with game information.
Also added in game error messages for when the player tries an action
that cannot be accomplished (ex. not enough resources to build a
building) to provide feedback.